chore: extract page bindings controller from `UtilityScript`
